move pytest fixtures outside library
so that building the docs doesn't require installing pytest
